week4/TodoApp/requirements.txt
fastapi==0.101.1
passlib==1.7.4
pydantic==1.10.5
python_jose==3.3.0
SQLAlchemy==2.0.7
starlette==0.31.0
